#86998medm

+++ packages/server/src/database/repositories/file.repository.mts
@@ -1,7 +1,8 @@
 import { inject, injectable } from "inversify";
-import { QueryRunner } from "typeorm";
+import { FindManyOptions, FindOneOptions, QueryRunner } from "typeorm";
 import { DbService } from "@way-to-bot/server/services/db.service.mjs";
 import { FileEntity } from "@way-to-bot/server/database/entities/file.entity.mjs";
+import { NotFoundError } from "@way-to-bot/server/common/errors/not-found.error.mjs";
 
 @injectable()
 export class FileRepository {
@@ -14,12 +15,12 @@ export class FileRepository {
     return this._dbService.dataSource.getRepository(FileEntity);
   }
 
-  getOneById(id: number, queryRunner?: QueryRunner) {
-    return this.getRepository(queryRunner).findOne({ where: { id } });
+  getOne(options: FindOneOptions<FileEntity>, queryRunner?: QueryRunner) {
+    return this.getRepository(queryRunner).findOne(options);
   }
 
-  getAll(queryRunner?: QueryRunner) {
-    return this.getRepository(queryRunner).find();
+  getMany(options?: FindManyOptions<FileEntity>, queryRunner?: QueryRunner) {
+    return this.getRepository(queryRunner).find(options);
   }
 
   async create(
@@ -29,11 +30,20 @@ export class FileRepository {
     const repo = this.getRepository(queryRunner);
     const newFile = repo.create(payload);
     const savedFile = await repo.save(newFile);
-    return this.getOneById(savedFile.id, queryRunner);
+    return this.getOne({ where: { id: savedFile.id } }, queryRunner);
   }
 
   async delete(id: number, queryRunner?: QueryRunner) {
     const repo = this.getRepository(queryRunner);
+    const existingFile = await this.getOne(
+      { where: { id: id }, relations: undefined },
+      queryRunner,
+    );
+
+    if (!existingFile) {
+      throw new NotFoundError(`File with id ${id} not found`);
+    }
+
     const result = await repo.delete(id);
     return result.affected === 1;
   }
